added conditions for specific elements to appear

+++ entity/index.php
@@ -31,19 +31,22 @@ $title = 'Recipes';
 			<div class="d-flex ms-3 buttons">
 				<div id="create">
 					<a href="create.php" id="create-btn" class="btn btn-primary">
-						Create Recipes
+						Create a Recipe
 					</a>
 				</div>
 
 				<?php
 				//Displays admin buttons if user has admin status
 				if (isset($_SESSION['admin'])) {
-					?>
-					<div>
-						<a href="../admin/delete-all.php?target=recipe" class="btn btn-danger">
-							Delete All Recipes
-						</a>
-					</div>
+					if (count($recipes) > 0) {
+						?>
+
+						<div>
+							<a href="../admin/delete-all.php?target=recipe" class="btn btn-danger">
+								Delete All Recipes
+							</a>
+						</div>
+					<?php } ?>
 
 					<div>
 						<a href="../admin/admin.php" class="btn btn-secondary manage-users">Manage Users</a>
@@ -52,9 +55,21 @@ $title = 'Recipes';
 			</div>
 		<?php } ?>
 
-		<div id="content">
-			<?php displayCards($recipes); ?>
-		</div>
+		<?php if (count($recipes) > 0) { ?>
+			<div id="content">
+				<?php displayCards($recipes); ?>
+			</div>
+		<?php } else {
+			?>
+			<div class="no-recipes">
+				<h1>No recipes to display.</h1>
+
+				<?php if (!isset($_SESSION['signedIn'])) { ?>
+					<p><a href="../auth/signin.php">Sign in</a> or <a href="../auth/signup.php">sign up</a> to add the first recipe.
+					</p>
+				<?php } ?>
+			</div>
+		<?php } ?>
 	</main>
 </body>
